backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1

# Supabase
supabase==2.3.4
python-jose[cryptography]==3.3.0

# OpenAI
openai==1.12.0

# Google APIs
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0

# VAPI
httpx==0.26.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Vector embeddings
numpy==1.26.4
